EES-3652 fixes for decimals, negatives and default colour

+++ src/GovUk.Education.ExploreEducationStatistics.Common/Model/Chart/ChartCustomDataGroup.cs
@@ -2,7 +2,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Common.Model.Chart;
 
 public class ChartCustomDataGroup
 {
-    public int Min;
-    public int Max; 
+  public decimal Min;
+  public decimal Max;
 }
